+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Changed
 
 - tzkt: Signatures of `[get/iter]_similar_contracts` and `[get/iter]_originated_contracts` methods have changed.
+- config: Removed `# dipdup: ignore`
+- config: Removed deprecated `similar_to.address` filter
+- cli: --file allias changed from -f to -o, because everywhere else -f means --force
+- cli: `init --overwrite-types` flag renamed to `--force` and now affects ABIs
 
 ### Removed